[feature] modify search by id
