+++ app/models.py
@@ -1,26 +1,2 @@
 #MODELs
-
-class User(db.Model):
-    """
-    Table Schema
-    """
-    __tablename__ = "users"
-
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    name = db.Column(db.String(50))
-    email = db.Column(db.String(50))
-    password = db.Column(db.String(50))
-
-
-class Event(db.Model):
-    """
-    Table Schema
-    """
-
-    __tablename__ = "events"
-
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    title = db.Column(db.String(50))
-    category = db.Column(db.String(50))
-    location = db.Column(db.String(50))
-    description = db.Column(db.String)
+from api import db
